add first draft of MySQLPDO_Manage
